Add files via upload
A Christmas theme project

develop an application to help a user to keep track of potential Christmas gifts.  The application should accept the recipient name, gift idea and a URL for that gift from the user and store them in an appropriate data structure (array or ArrayList of objects).  The application should have a full swing user GUI and allow the user to add gift ideas, delete gift ideas, display all gift ideas, and search for a gift by recipient name. Make sure you can Save and Load the data even when the program stops running.